requirements.txt
# Arkalia Quest - Dépendances Python
# Version 2.0.0

# Framework web
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

# Tests et qualité de code
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-html==4.1.1
coverage==7.3.2
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Traitement d'images
Pillow==10.1.0

# Utilitaires
python-dotenv==1.0.0
click==8.1.7
colorama==0.4.6
requests==2.31.0

# Sécurité
cryptography==41.0.7

# Performance et monitoring
psutil==5.9.6

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Développement
pre-commit==3.5.0

# ===== NOUVELLES DÉPENDANCES POUR VERSION 3.0 =====

# WebSockets pour défis temps réel
Flask-SocketIO==5.3.6
python-socketio==5.9.0
eventlet==0.33.3

# IA et Machine Learning
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2

# Base de données SQLite (déjà intégré dans Python)
# sqlite3 est inclus dans Python standard
